* fix for is_mane_server error during migration
MMS-3 Главный сервер в кластере

+++ plugins/wpstudio/mms/updates/builder_table_update_wpstudio_mms_servers_4.php
@@ -8,17 +8,23 @@ class BuilderTableUpdateWpstudioMmsServers4 extends Migration
 {
     public function up()
     {
-        Schema::table('wpstudio_mms_servers', function (Blueprint $table) {
-            $table->boolean('is_main_server')->nullable();
+        if (!Schema::hasColumn('wpstudio_mms_servers', 'is_main_server')){
+            Schema::table('wpstudio_mms_servers', function (Blueprint $table) {
+                $table->boolean('is_main_server')->nullable();
 
-            $table->unique(['cluster_id', 'is_main_server'], 'unique_cluster_main_server');
-        });
+                $table->unique(['cluster_id', 'is_main_server'], 'unique_cluster_main_server');
+            });
+        }
     }
     
     public function down()
     {
-        Schema::table('wpstudio_mms_servers', function (Blueprint $table) {
-            $table->dropColumn('is_main_server');
-        });
+        if (Schema::hasColumn('wpstudio_mms_servers', 'is_main_server')) {
+            Schema::table('wpstudio_mms_servers', function (Blueprint $table) {
+                $table->dropUnique('unique_cluster_main_server');
+
+                $table->dropColumn('is_main_server');
+            });
+        }
     }
 }
